rename random baselines and extend artifitical datasets
